xserver putty mesa fontconfig libX11 libXmu pixman xkeyboard-config git update May 20 2018
xserver          commit 7437b6dbdee050f8ebb3a79b9077d051c91880c3
xkeyboard-config commit d349d52ce409b814bd59ec3beadfd6263a7582fa
libX11           commit 9c5845ff0dc080ff10bd68af4fc40fcd805728ca
pixman           commit ddf42d627c724ef5337ea86e1ac7ef24a5df8f37
libXmu           commit 5bd4eb821bd9e88c3a131de9a8b781f5184e22b1
fontconfig       commit f098adac54ab86b75a38f2d23fa706a1348f55ba
mesa             commit 549e54270ba3a519b46a1fbffa4aa6b628a052d3
putty            commit 7d0ade7eac694d63381054d38d3eb4f199c350d7

+++ fontconfig/test/run-test.sh
@@ -203,4 +203,29 @@ fi
 rm -rf $TESTTMPDIR out1 out2 xxx bind-fonts.conf
 fi
 
+dotest "sysroot option"
+prep
+mkdir -p $MyPWD/sysroot/$FONTDIR
+mkdir -p $MyPWD/sysroot/$CACHEDIR
+cp $FONT1 $MyPWD/sysroot/$FONTDIR
+cp $MyPWD/fonts.conf $MyPWD/sysroot/$MyPWD/fonts.conf
+$FCCACHE -y $MyPWD/sysroot
+stat $MyPWD/sysroot/$FONTDIR/.uuid
+if test $? != 0; then
+  echo "*** Test failed: $TEST"
+  exit 1
+fi
+
+dotest "creating uuid-based cache file on sysroot"
+uuid=`cat $MyPWD/sysroot/$FONTDIR/.uuid`
+ls $MyPWD/sysroot/$CACHEDIR/$uuid*
+if [ $? != 0 ]; then
+  echo "*** Test failed: $TEST"
+  echo "No cache for $uuid"
+  ls $MyPWD/sysroot/$CACHEDIR
+  exit 1
+fi
+
+rm -rf $MyPWD/sysroot
+
 rm -rf $FONTDIR $CACHEFILE $CACHEDIR $FONTCONFIG_FILE out
